General fixes to POWER CAS atomic routines
This patch ported the big-endian (BE) version of the POWER CAS assembly
routines, as well as fixes for some errors found on the little-endian (LE)
ones, basically by calling the internal GCC macros directly (and not relying
on Intel TBB to do so). This seems to fix some of the random errors found on
previous LE test executions. This also provides some syntax cleanup on the
power specific file and atomic.h.